Code cleanup & modernization

+++ src/main/java/com/cburch/logisim/std/arith/FpToInt.java
@@ -29,10 +29,10 @@ import java.awt.Color;
 
 public class FpToInt extends InstanceFactory {
   /**
-   * Unique identifier of the tool, used as reference in project files.
-   * Do NOT change as it will prevent project files from loading.
+   * Unique identifier of the tool, used as reference in project files. Do NOT change as it will
+   * prevent project files from loading.
    *
-   * Identifier value must MUST be unique string among all tools.
+   * <p>Identifier value must MUST be unique string among all tools.
    */
   public static final String _ID = "FPToInt";
 
@@ -100,7 +100,7 @@ public class FpToInt extends InstanceFactory {
 
     if (roundMode.getValue().equals("ceil")) out_val = (long) Math.ceil(a_val);
     else if (roundMode.getValue().equals("floor")) out_val = (long) Math.floor(a_val);
-    else if (roundMode.getValue().equals("round")) out_val = (long) Math.round(a_val);
+    else if (roundMode.getValue().equals("round")) out_val = Math.round(a_val);
     else out_val = (long) a_val;
 
     final var out = Value.createKnown(dataWidthOut, out_val);
